% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfs_ productivity_downloader.R
\name{download_conus_productivity}
\alias{download_conus_productivity}
\title{USFS CONUS Site Productivity 2018 Downloader}
\usage{
download_conus_productivity(
  aoi,
  output_path = "conus_productivity_output.tif",
  max_tile_size = 20000,
  mask = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{aoi}{the Area of Interest that you would like to download the site productivity raster for}

\item{output_path}{A character string specifying the file path for the output raster (GeoTIFF). Default is \code{"esri_rest_output.tif"}.}

\item{max_tile_size}{Integer specifying the maximum size (in pixels) of each downloaded tile. Default is \code{20000}.}

\item{mask}{Logical; if \code{TRUE}, the merged raster will be masked to the AOI boundary. Default is \code{TRUE}.}

\item{quiet}{Logical; if \code{TRUE}, suppresses progress messages. Default is \code{FALSE}.}
}
\value{
A \code{SpatRaster} object representing the merged (and optionally masked) raster. The final raster is also saved to \code{output_path}.
}
\description{
This downloads the USFS CONUS Site Productivity layer from the USFS ArcGIS REST API server.
The site productivity raster is a 30x30 m grid that ranks the site from 1 - 7.

From the USFS:
This image service was developed using data from over 213,000 national forest inventory plots measured during the period 2014-2018 from the USFS Forest Inventory and Analysis (FIA) program, in conjunction with other auxiliary information. Roughly 4,900 Landsat 8 OLI scenes, collected during the same time period, were processed to extract information about vegetation phenology. This information, along with climatic and topographic raster data, were used in an ecological ordination model of tree species. The model produced a feature space of ecological gradients that was then used to impute FIA plots to pixels. The plots imputed to each pixel were then used to assign values of site productivity.

Code Description
1 = 225+ cubic feet/acre/year.
2 = 165-224 cubic feet/acre/year.
5 = 50-84 cubic feet/acre/year.
6 = 20-49 cubic feet/acre/year.
7 = 0-19 cubic feet/acre/year.

More information and citation of the data along with methodology can be found here : https://www.arcgis.com/home/item.html?id=b1b10e7890ea4116b863ae790d9b718c
}
\details{
The raster is 30x30 m in resolution. The REST API allows for a max tile size of 117545 x 214620
}
\examples{
\dontrun{
library(sf)

# Define a sample AOI as a bounding box polygon
aoi <- st_as_sfc(st_bbox(c(xmin = -100, ymin = 40, xmax = -99, ymax = 41), crs = st_crs(4326)))

# Download the raster from an ESRI REST image service
raster <- ddownload_conus_productivity(aoi,
                               output_path = "conus_productivity.tif")
}

}
